requirements.txt
# Federated Learning
flwr==1.7.0
flower==2.0.1

# Machine Learning
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Explainability
shap==0.44.1

# LLM Integration
google-generativeai==0.3.2

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Data Handling
requests==2.31.0
ucimlrepo==0.0.3

# Visualization (for SHAP)
matplotlib==3.8.2
plotly==5.18.0

# Utilities
tqdm==4.66.1
